pyproject.toml
[project]
name = "score"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "bertopic>=0.16.4",
    "docx>=0.2.4",
    "fastapi[standard]>=0.115.11",
    "fitz>=0.0.1.dev2",
    "gensim>=4.3.3",
    "ipykernel>=6.29.5",
    "keybert>=0.9.0",
    "nltk>=3.9.1",
    "pdfplumber>=0.11.5",
    "pillow>=11.1.0",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.8.1",
    "pypdf2>=3.0.1",
    "pytesseract>=0.3.13",
    "python-multipart>=0.0.20",
    "spacy>=3.8.4",
    "sqlalchemy>=2.0.39",
    "streamlit>=1.43.2",
    "supabase>=2.13.0",
    "torch>=2.6.0",
    "torchvision>=0.21.0",
    "transformers>=4.49.0",
    "yake>=0.4.8",
]
